+++ integration-test-query-logger/src/test/java/com/ericsson/bss/cassandra/ecaudit/integration/querylogger/ITQueryLogger.java
@@ -83,7 +83,7 @@ public class ITQueryLogger
     }
 
     @AfterClass
-    public static void afterClass() 
+    public static void afterClass()
     {
         session.close();
     }