+++ ecaudit/src/main/java/com/ericsson/bss/cassandra/ecaudit/facade/Auditor.java
@@ -53,6 +53,8 @@ public interface Auditor
      */
     boolean shouldLogFailedBatchSummary();
 
+    boolean shouldLogPrepareStatements();
+
     /**
      * Sets the log timing strategy to use.
      *