+++ ecaudit/src/main/java/com/ericsson/bss/cassandra/ecaudit/filter/AuditFilter.java
@@ -37,4 +37,6 @@ public interface AuditFilter {
      * For example, use this method to create any required keyspaces/tables.
      */
     void setup();
+
+    boolean shouldLogPrepareStatements();
 }